+++ example/integration_test/downloader_integration_test.dart
@@ -154,6 +154,14 @@ void main() {
           baseDirectory: BaseDirectory.applicationSupport);
       path = join((await getApplicationSupportDirectory()).path, task.filename);
       await enqueueAndFileExists(path);
+      // applicationLibrary directory
+      task = DownloadTask(
+          url: workingUrl,
+          filename: defaultFilename,
+          baseDirectory: BaseDirectory.applicationLibrary);
+      path = await task.filePath();
+      await enqueueAndFileExists(path);
+
       // test url with encoded parameter
       task = DownloadTask(
           url: getTestUrl,
@@ -1649,6 +1657,9 @@ void main() {
   });
 
   group('Notifications', () {
+    // NOTE: notifications are difficult to test in an integration test, so
+    // passing tests in this group is not sufficient evidence that they
+    // are working properly
     testWidgets('NotificationConfig', (widgetTester) async {
       FileDownloader()
           .configureNotification(running: TaskNotification('Title', 'Body'));
@@ -1662,6 +1673,92 @@ void main() {
           allowPause: true);
       expect(await FileDownloader().enqueue(task), equals(true));
       await statusCallbackCompleter.future;
+    });
+  });
+
+  group('Shared storage', () {
+    test('move task to shared storage', () async {
+      var filePath = await task.filePath();
+      await FileDownloader().download(task);
+      final path = await FileDownloader()
+          .moveToSharedStorage(task, SharedStorage.downloads);
+      print('Path in downloads is $path');
+      expect(path, isNotNull);
+      expect(File(filePath).existsSync(), isFalse);
+      expect(File(path!).existsSync(), isTrue);
+      File(path).deleteSync();
+    });
+
+    test('move task to shared storage with directory', () async {
+      var filePath = await task.filePath();
+      await FileDownloader().download(task);
+      final path = await FileDownloader().moveToSharedStorage(
+          task, SharedStorage.downloads,
+          directory: 'subdirectory');
+      print('Path in downloads is $path');
+      expect(path, isNotNull);
+      expect(File(filePath).existsSync(), isFalse);
+      expect(File(path!).existsSync(), isTrue);
+      File(path).deleteSync();
+      expect(dirname(path).endsWith('subdirectory'), isTrue);
+      Directory(dirname(path)).deleteSync();
+    });
+
+    test('try to move text file to images -> error', () async {
+      if (Platform.isAndroid) {
+        var filePath = await task.filePath();
+        await FileDownloader().download(task);
+        final path = await FileDownloader()
+            .moveToSharedStorage(task, SharedStorage.images);
+        expect(path, isNull);
+        expect(File(filePath).existsSync(), isTrue);
+      }
+    });
+
+    test('move file to shared storage - all types', () async {
+      for (var destination in SharedStorage.values) {
+        await FileDownloader().download(task);
+        var filePath = await task.filePath();
+        expect(File(filePath).existsSync(), isTrue);
+        // rename the file extension to accommodate requirement for shared
+        // storage (e.g. an .html file cannot be stored in 'images')
+        switch (destination) {
+          case SharedStorage.images:
+            final newFilePath = filePath.replaceFirst('.html', '.jpg');
+            await File(filePath).rename(newFilePath);
+            filePath = newFilePath;
+            break;
+          case SharedStorage.video:
+            final newFilePath = filePath.replaceFirst('.html', '.mp4');
+            await File(filePath).rename(newFilePath);
+            filePath = newFilePath;
+            break;
+          case SharedStorage.audio:
+            final newFilePath = filePath.replaceFirst('.html', '.mp3');
+            await File(filePath).rename(newFilePath);
+            filePath = newFilePath;
+            break;
+          default:
+            break;
+        }
+        final path = await FileDownloader()
+            .moveFileToSharedStorage(filePath, destination);
+        print('Path in shared storage for $destination is $path');
+        if (Platform.isAndroid ||
+            destination == SharedStorage.downloads ||
+            (Platform.isIOS &&
+                destination != SharedStorage.files &&
+                destination != SharedStorage.external)) {
+          expect(path, isNotNull);
+          expect(File(filePath).existsSync(), isFalse);
+          expect(File(path!).existsSync(), isTrue);
+          File(path).deleteSync();
+        } else {
+          // otherwise expect null
+          expect(path, isNull);
+          File(filePath).deleteSync();
+        }
+      }
     });
   });
 }